requirements.txt
# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.15.0
webdriver-manager>=4.0.1

# OCR and document processing
boto3>=1.29.0
azure-ai-formrecognizer>=3.3.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.1

# Data processing
pandas>=2.1.0
numpy>=1.25.0
openpyxl>=3.1.2

# Database support
sqlalchemy>=2.0.0

# Configuration and environment
python-decouple>=3.8
pydantic>=2.4.0
pydantic-settings>=2.0.3

# Logging and monitoring
structlog>=23.1.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1

# Code quality
black>=23.9.1
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Utilities
tqdm>=4.66.0
click>=8.1.7
python-dateutil>=2.8.2
validators>=0.22.0

# Security
cryptography>=41.0.0
